requirements.txt
# 마음정류장 채널 필수 의존성
Pillow>=10.4.0
google-api-python-client>=2.137.0
google-auth-oauthlib>=1.2.1
google-auth>=2.34.0
moviepy>=1.0.3
openai>=1.43.0
pydub>=0.25.1
python-dotenv>=1.0.1
srt>=3.5.3
schedule>=1.2.0

# 선택적 의존성 (고급 기능용)
# numpy>=1.24.0  # MoviePy 성능 향상
# imageio-ffmpeg>=0.4.8  # 고품질 비디오 인코딩
